:white_check_mark: Separate tests and fix errors
