+++ backend/zane_api/tests/__init__.py
@@ -1,0 +1,3 @@
+from .auth import *
+from .docker import *
+from .project import *
