+++ backend/zane_api/views/docker.py
@@ -52,18 +52,7 @@ class DockerImageSearchView(APIView):
         if form.is_valid():
             params = form.data
             result = DockerService.search_registry(term=params["q"])
-
-            images_to_return = []
-            for image in result:
-                api_image_result = {}
-                if image["is_official"]:
-                    api_image_result["full_image"] = f'library/{image["name"]}:latest'
-                else:
-                    api_image_result["full_image"] = f'{image["name"]}:latest'
-                api_image_result["description"] = image["description"]
-                images_to_return.append(api_image_result)
-
-            response = self.serializer_class({"images": images_to_return})
+            response = self.serializer_class({"images": result})
             return Response(response.data, status=status.HTTP_200_OK)
 
         return Response(
